improved ball/paddle bouncing
The speed of the ball isn't constant anymore, but depends on a base
speed (fixed through eventual levels/bonuses later (who knows)) and on
its angle. I find it improves gameplay somewhat.
I've done away with all the sf:: prefixing that was getting annoying,
and fixed a few things. Not much change.