add suppport for dynamically-sized maps
Map sizes are no longer hardcoded, and on map change the window is resized
accordingly. Only the brick size and the possible locations for bricks are
still hardcoded, and I plan to de-hardcode the latter.
Changes to the Makefile are supposed to force a (partial) rebuild even when
only header files have changed. I've had weird segfaults because of this.