vastly improved collisions
There are still quirks now and again but the game is now a lot more
playable than before, as the bounces are more predictable. A smaller
ball should reduce errors, I'll have to test that.
Oh, and hate on my use of `goto` all you want, it works juust fine.