add initial support for different brick types
And with it, at last, sprite support for bricks. I was unsure how to
implement the mapping between brick types and their texture, but I ended
up hardcoding it. Another way was to move maps to json and hardcode the
paths per map. It clearly wasn't worth the complexity and added
dependencies. Plus generating maps would've gotten harder.