special bricks now drop bonuses
Once hit, a special brick now falls straight down and has to be touched
by the paddle to trigger the bonus. Currently, the following bonuses are
implemented: expand the paddle, and gain one life (you don't lose if you
drop the ball).
Another notable change is the switch from integer, 'grid' coordinates
for the bricks to screen coordinates. It allows for moving a brick
smoothly, and is more consistent with how other objects function.