implement `new ball` bonus
When the bonus is caught, the latest ball to have appeared is
duplicated, with the new version having an inverted speed along the
x-axis. I think (I didn't check) the standard behavior is for the new
ball to appear on the player's paddle, but the behavior I chose is both
easier to implement (given the way things are) and slightly more cool.