improve ball look, fix speed bug
Away with magenta. I probably picked that color in the first place just
to see what it was, and it stuck. I stick with a plain circle as there
just aren't enough pixels to make something else look tolerable, and it
fits the retro-ish feel.
As for the bug fixed, it was a division by zero during speed
normalization right after the ball was unstuck. The thing is, this bug
has been present since the second commit of this repository, and yet it
used to work perfectly... the joys of undefined behaviors.