add level editor
Here's an overview of how it works. First it's a Godot 3 project, and
the Godot engine has to be downloaded separately (it's about 30MB).
The project consists of two main scenes: one used to generate the tile
set from textures and the color palette, and the other to actually
design levels and export them to breaker-readable files.
Here's how to use those components:
 - Creating a level is as easy as adding a new TileMap node as a child of
the `breaker_maps` node, setting the tileset/grid size (or simply
copying from an old map), drawing the level and pressing F5. The file
will have been created under the `maps` folder of the Godot project.
 - Modifying the tile set requires adding the necessary textures to the
`textures` directory within the project, which will be used only in the
editor, if needed editing the `palette.list` file in the main breaker
folder and the type_to_id dict in `tileset.gd`, deleting the children
of the `tileset` node and closing and reopening the `tileset.tscn` scene.
Finally, click Scene->Convert to->TileSet and overwrite `tileset.tres`.
This process is unneeded most of the time, and many steps are optional
depending on what is being changed.